requirements.txt
psycopg2-binary
fastapi
uvicorn
pydantic
python-multipart
requests
pytest
httpx
sqlmodel
google-generativeai
python-dotenv
